+++ src/spatialdata_io/experimental/__init__.py
@@ -1,0 +1,6 @@
+from spatialdata_io.converters.legacy_anndata import (
+    from_legacy_anndata,
+    to_legacy_anndata,
+)
+
+__all__ = ["from_legacy_anndata", "to_legacy_anndata"]