+++ CHANGELOG.md
@@ -15,7 +15,8 @@ and this project adheres to [Semantic Versioning][].
 -   (MCMICRO) support for TMAs (such as the data of exemplar-002)
 -   (Xenium) support for post-xenium aligned images (IF, HE)
 -   (Xenium) reader for the selection coordinates file from the Xenium Explorer
--   DBiT-seq reader
+-   (DBiT-seq) reader
+-   converter functions `experimental.to_legacy_anndata()` and `experimental.from_legacy_anndata()`
 -   (Visium) support for raw reads (capture locations not under tissue)
 
 ### Fixed